requirements.txt
django==1.10.0
mysqlclient==1.3.7
djangorestframework==3.4.6
django-rest-auth[with_social]==0.8.1
